* cluster compile fix (gonna start working again on this when i'm feeling better) * numerous crash+performance fixes * possible fix for netcode disconnection issue *this code needs a rewrite*
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2473 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/WorldSocket.h
@@ -241,7 +241,7 @@ public:
 	inline string GetRemoteIP() { return string(inet_ntoa(m_address.sin_addr)); }
 	inline uint32 GetRemotePort() { return ntohs(m_address.sin_port); }
 
-	inline void SendPacket(WorldPacket* packet) { if(!packet) return; OutPacket(packet->GetOpcode(), packet->size(), (packet->size() ? (const void*)packet->contents() : NULL)); }
+	inline void SendPacket(WorldPacket* packet) { if(!packet) return; OutPacket(packet->GetOpcode(), (uint16)packet->size(), (packet->size() ? (const void*)packet->contents() : NULL)); }
 	inline void SendPacket(StackBufferBase * packet) { if(!packet) return; OutPacket(packet->GetOpcode(), packet->GetSize(), (packet->GetSize() ? (const void*)packet->GetBufferPointer() : NULL)); }
 	void __fastcall OutPacket(uint16 opcode, uint16 len, const void* data);
 	inline uint32 GetSessionId() { return m_sessionId; }
